♻️ Refactor into constants and util
