+++ index.ts
@@ -1,14 +1,10 @@
 import yargs from "yargs"
-import notifier from "node-notifier"
-
-// Default values
-const DEFAULT_ROLES_4_PLAYERS = ["Driver", "Navigator", "Facilitator", "Scribe"]
-const DEFAULT_ROLES_3_PLAYERS = ["Driver", "Navigator", "Facilitator/Scribe"]
-const ERROR_MSG_PLAYERS =
-  "Error: There must be exactly 3 or 4 players specified."
-
-const DEFAULT_TOTAL_DURATION = 120 // Default total duration in minutes
-const DEFAULT_INTERVAL_DURATION = 15 // Default interval duration in minutes
+import {
+  DEFAULT_TOTAL_DURATION,
+  DEFAULT_INTERVAL_DURATION,
+  ERROR_MSG_PLAYERS,
+} from "./constants"
+import { getRoles, isValidPlayerCount, rotatePlayers } from "./util"
 
 // Parse command line arguments
 const argv = yargs
@@ -47,67 +43,3 @@ const numAlerts = Math.floor(totalDuration / intervalDuration)
 
 // Start rotation
 rotatePlayers(players, roles, numAlerts, intervalDuration, totalDuration)
-
-// Function to validate the number of players
-function isValidPlayerCount(playerCount: number) {
-  return playerCount === 3 || playerCount === 4
-}
-
-// Function to get roles based on the number of players
-function getRoles(playerCount: number) {
-  return playerCount === 4 ? DEFAULT_ROLES_4_PLAYERS : DEFAULT_ROLES_3_PLAYERS
-}
-
-// Function to rotate players and roles and display alerts
-function rotatePlayers(
-  players: string[],
-  roles: string[],
-  numAlerts: number,
-  intervalDuration: number,
-  totalDuration: number,
-) {
-  let index = 0
-
-  for (let i = 0; i < numAlerts; i++) {
-    setTimeout(
-      () => {
-        const rotation = createRotation(players, roles, index)
-        console.log(`[${i + 1}] ${rotation}`)
-        index = (index + 1) % players.length
-
-        sendNotification(`Rotation Alert ${i + 1}`, rotation)
-      },
-      i * intervalDuration * 60 * 1000,
-    ) // i * interval duration in milliseconds
-  }
-
-  // Add a final setTimeout to keep the script running until the last alert is done
-  setTimeout(
-    () => {
-      console.log("Rotation session completed.")
-      process.exit(0) // Optionally exit the process after completion
-    },
-    totalDuration * 60 * 1000,
-  )
-}
-
-// Function to create a rotation string
-function createRotation(
-  players: string[],
-  roles: string[],
-  startIndex: number,
-) {
-  return roles
-    .map((role, j) => `${players[(startIndex + j) % players.length]} - ${role}`)
-    .join(", ")
-}
-
-// Function to send desktop notification
-function sendNotification(title: string, message: string) {
-  notifier.notify({
-    title,
-    message,
-    sound: true,
-    wait: false,
-  })
-}
